+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- added: Support for blockbook server connections with "%{key-name}" key parameters
 - fixed: Ensure server scoring is done on all server tasks
 - changed: Refactor pickNextTask to implement generators to replace Deferred promises